list double click, complete SQL

+++ mainwindow.cpp
@@ -26,6 +26,7 @@ void MainWindow::on_cal_selectionChanged()
 void MainWindow::on_addBtn_clicked() // add item
 {
     choose *ch = new choose;
+    ch->isAdd = true; // add item
     connect(ch, SIGNAL(sendData(QString)), this, SLOT(receiveData(QString)));
     ch->show();
 }
@@ -33,18 +34,38 @@ void MainWindow::on_addBtn_clicked() // add item
 void MainWindow::on_delBtn_clicked() // delete item
 {
     int i = ui->list->currentRow();
+    if(i != -1)
+    {
+        QSqlQuery query;
+        QString title = ui->list->item(i)->text();
+        QString ss = "delete from List where title = '" + title + "'";
+        query.exec(ss);
 
-    QSqlQuery query;
-    QString title = ui->list->item(i)->text();
-    QString ss = "delete from List where title = '" + title + "'";
-    query.exec(ss);
+        delete ui->list->item(i);
+    }
 
-    delete ui->list->item(i);
 }
 
 void MainWindow::on_list_doubleClicked(const QModelIndex &index)
 {
-    // map
+    int i = ui->list->currentRow();
+    QString title = ui->list->item(i)->text();
+    QSqlQuery query;
+    QString ss = "SELECT * FROM List where title = '" + title + "'";
+    query.exec(ss);
+    query.first();
+
+    QString title2 = query.value(0).toString();
+    QTime startTime = query.value(1).toTime();
+    QDateTime endDateTime = query.value(2).toDateTime();
+    QString curLocation = query.value(3).toString();
+    QString destination = query.value(4).toString();
+    QString note = query.value(5).toString();
+
+    choose *ch = new choose;
+    ch->setting(title2, startTime, endDateTime, curLocation, destination, note);
+    ch->isAdd = false; // edit item
+    ch->show();
 }
 
 void MainWindow::receiveData(QString data)
@@ -90,11 +111,6 @@ void MainWindow::on_check_clicked()
         qDebug() << query.value(5).toString();
         qDebug() << "----data finish----";
     }
-}
-
-void MainWindow::get_title(QString input)
-{
-    title = input;
 }
 
 void MainWindow::on_re_clicked()
